Fixed error

+++ include/solverTile.h
@@ -7,7 +7,7 @@
 struct SolverTile
 {
     std::vector<SolverTile*> adjSolverTiles;
-    float bombProbability;
+    double bombProbability;
     uint16_t x;
     uint16_t y;
     uint16_t adjBombsAmount;
